Get itens in a list as a new Object
